backend/app/services/ai_orchestrator.py
import asyncio
import os
import google.generativeai as genai

# Configure the Gemini API key
# Make sure to set the GOOGLE_API_KEY environment variable
genai.configure(api_key=os.environ["GOOGLE_API_KEY"])

async def process_media_chunk_with_gemini(chunk: bytes):
    """
    This is a placeholder function to simulate processing a media chunk with Gemini.
    In a real application, this would involve sending the chunk to the Gemini API
    and receiving a response.
    """
    print(f"Processing chunk of size {len(chunk)} with Gemini")

    # This is where you would interact with the Gemini API.
    # The following code is a conceptual example and will not run without a
    # proper implementation of the Gemini streaming API.

    # model = genai.GenerativeModel('gemini-pro-vision')
    # response = await model.generate_content_async(
    #     contents=[
    #         {"parts": [{"inline_data": {"mime_type": "video/webm", "data": chunk}}]},
    #         {"parts": [{"text": "What is happening in this video?"}]}
    #     ],
    #     stream=True
    # )
    #
    # async for chunk in response:
    #     print(chunk.text)

    await asyncio.sleep(1)  # Simulate a delay for AI processing
    return {
        "text": "This is a response from the Gemini placeholder.",
        "audio_url": "https://example.com/gemini_response.mp3",
        "video_url": "https://example.com/gemini_response.mp4",
    }
