backend/requirements.txt
fastapi
uvicorn
python-dotenv
websockets
livekit
google-generativeai

fastapi
uvicorn